Move files into folders for organization
