+++ rps/ascii_art.rb
