+++ toh/ascii_art.rb
