+++ warmups/enumerable_methods.rb
